Move logger.js and modules.js into utils
